resume_screening/requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.0
python-docx>=1.0.0
spacy>=3.7.0
scikit-learn>=1.3.0
pandas>=2.0.0
plotly>=5.17.0
Flask>=3.0.0
gunicorn>=21.2.0

